+++ apps/cms-v2/src/modules/providers/cms-provider-meta.ts
@@ -1,0 +1,9 @@
+import { ReactElement } from "react";
+
+export interface CMSProviderMeta {
+  type: string;
+  displayName: string;
+  logoUrl: string;
+  description: string;
+  formSideInfo?: ReactElement;
+}
