+++ apps/cms-v2/src/modules/providers/contentful/contentful.tsx
@@ -1,7 +1,7 @@
 import { Text } from "@saleor/macaw-ui/next";
 
 import logo from "./contentful-logo.svg";
-import { CMSProviderMeta } from "../providers-resolver";
+import { CMSProviderMeta } from "../cms-provider-meta";
 
 export const Contentful = {
   formSideInfo: <Text>App will save each variant with the same ID as variant ID.</Text>,
